+++ doc/ReleaseNotes.md
@@ -1,6 +1,14 @@
 Release Notes for Version 13.0
 =============================
 
+Build 001
+-------
+
+New Features:
+
+* The getString and getStringJS methods of the ResBundle class both now take either a string or an array parameter
+to translate. When an array is given, every string in the array is translated.
+
 Build 000
 -------